syncing to separate Sync component
